redwood: overlay: Define thresholds for switching refresh rate
The display uses different gamma curves for different refresh rates.
It's hard for panel vendor to tune the curves to have exact same brightness
for different refresh rate (lazy). To avoid flicker at darker environment,
set threshold for 120Hz, so 120Hz will be default in such conditions.

Also, increasing rr switching timers, even if they made the CRT bug to be reduced,
slowly scrolling on GMaps while Smooth Display is enabled, led to a weird gamma
switching to a green tint, so we need to tune that too for avoiding that shifting.